Chinh sua giao dien

+++ sql_script/script.sql
@@ -113,7 +113,7 @@ CREATE TABLE Shift (
     startTime TIME NOT NULL, 
     endTime TIME NOT NULL, 
     modifiedDate DATETIME NOT NULL, 
-    numberOfHour DOUBLE NOT NULL,
+    numberOfHour FLOAT NOT NULL,
     shiftDaysSchedule NVARCHAR(20) NOT NULL CHECK (shiftDaysSchedule IN ('MON_WEB_FRI', 'TUE_THU_SAT', 'SUNDAY')) 
 );
 GO
